First working implementation... complete.
A couple touchoups remain before the hackathon but nothing major. A
couple pieces of the implementation are pretty dirty but they still
work, e.g. the poller is nasty. Needs to be waiting on a signal. Also
program shutdown doesn't really work, need to figure out how to hook
into Tk to stop my other thread. But it still demos well.